requirements.txt
annotated-types==0.7.0
anyio==4.9.0
certifi==2025.8.3
click==8.2.1
colorama==0.4.6
deprecation==2.1.0
dnspython==2.7.0
email_validator==2.2.0
exceptiongroup==1.3.0
fastapi
gotrue==2.12.3
h11==0.16.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
hyperframe==6.1.0
idna==3.10
packaging==25.0
postgrest==1.1.1
pydantic

PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
realtime==2.6.0
six==1.17.0
sniffio==1.3.1
starlette==0.47.2
storage3==0.12.0
StrEnum==0.4.15
supabase
supafunc==0.10.1
typing-inspection==0.4.1
typing_extensions==4.14.1
uvicorn
websockets==15.0.1
sqlmodel==0.0.16
psycopg2-binary
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic-settings==2.0.3
stripe
